* commit '66a0fb80ee9120d646dc8a18f4fe67016d8c7528':
  Added bsquo entity decoding.
  Fixed -integerValueFromHex method for 64bits targets.
  Disambiguate count method.
  Fixed bundle selection.
  Switched back to old syntax when accessing array and dictionary items.
  Fixed shadowed variables.
  Fixed integer sign mismatch.

+++ Core/Source/DTHTMLParserNode.m
@@ -77,7 +77,7 @@
 	@synchronized(self)
 	{
 		// indent to the level
-		for (int i=0; i<indentLevel; i++)
+		for (NSUInteger i=0; i<indentLevel; i++)
 		{
 			[string appendString:@"   "];
 		}
@@ -109,7 +109,7 @@
 		}
 		
 		// indent to the level
-		for (int i=0; i<indentLevel; i++)
+		for (NSUInteger i=0; i<indentLevel; i++)
 		{
 			[string appendString:@"   "];
 		}
